drm/tidss: Add bandgap power control for OLDI LVDS buffers
Enable/disable the bandgap circuit for stabilizing the LVDS signals from
the OLDI TXes.

The common mode voltage of the LVDS buffers becomes random when the
bandgap reference is turned off. This causes uncertainity in the LVDS
Data and Clock signal outputs, making it behave differently under
different conditions and panel setups.

The bandgap reference must be powered on before using the OLDI IOs, to
keep the common voltage trimmed down to desired levels.

Signed-off-by: Nishanth Menon <nm@ti.com>
Signed-off-by: Aradhya Bhatia <a-bhatia1@ti.com>
Signed-off-by: Vignesh Raghavendra <vigneshr@ti.com>