arm64: dts: ti: Add J784S4 CPB OV5640 overlay
J784S4 EVM boards are not shipped with the OV5640 sensor module.
The module is just one of the possible sensors or capture boards
that can be interfaced. The OV5640 module is the module with which
the single camera capture on J784S4 is tested with. Add an overlay
for the inerfacing OV5640 module on the instance 2 (routed to the
connector CS2-EXP-AUX) and adds corresponding entry in Makefile.

Signed-off-by: Vaishnav Achath <vaishnav.a@ti.com>
Signed-off-by: Vignesh Raghavendra <vigneshr@ti.com>