media: ov5640: Handle delays when no reset_gpio set
Some RPi-compatible module manufacturers [1] don't expose the RESETB
gpio of the sensor directly through the connector. Instead wiring ~PWDN
as a proxy reset with appropriate delays.

In such cases, reset_gpio will not be available to the driver, but it
will still be toggled when the sensor is powered on, and thus we should
still honor the wait time of >= 5ms + 1ms + 20ms [2] before
attempting any i/o operations over SCCB.

[1] https://digilent.com/reference/_media/reference/add-ons/pcam-5c/pcam_5c_sch.pdf
[2] https://cdn.sparkfun.com/datasheets/Sensors/LightImaging/OV5640_datasheet.pdf

Signed-off-by: Jai Luthra <j-luthra@ti.com>