arm64: dts: ti: k3-am62a7-sk: Enable HDMI audio
Enable the HDMI bridge to act as a digital audio interface in the sound
node.

SK-AM62A has the same HDMI bridge as SK-AM62x, with muxed McASP1 lines
going to both codec and the bridge for audio output, thus same overlay
can be used to switch to HDMI audio.

Signed-off-by: Jai Luthra <j-luthra@ti.com>
Reviewed-by: Roger Quadros <rogerq@kernel.org>
Signed-off-by: Vignesh Raghavendra <vigneshr@ti.com>