HACK: drm/tidss: tidss_kms: mod param for am62 sk vs bb
So far tidss Driver only had support for display panels and bridges
that are compatible with connector framework. However, there are
drivers that aren't compatible with the same. TIDSS traditionally has
only supported the standard connector framework oriented drivers. Now,
using a module parameter that can be set to select between the bridge
drivers that need a bit of a "support".

This is still a hack, since this approach still applies the hack to
all bridges and panel instead of selectively applying the hack just to
the bridges that need this workaround.

Due to the nature of this workaround - the workaround is ONLY applied
for AM62x platform - leaving all other platforms as they stand.

For example: bootargs (via uEnv.txt or extlinux.conf) can be modified by
setting: tidss.no_connector_hack=0

Signed-off-by: Dhruva Gole <d-gole@ti.com>
Tested-by: Aradhya Bhatia <a-bhatia1@ti.com>
Cc: Nishanth Menon <nm@ti.com>
Signed-off-by: Nishanth Menon <nm@ti.com>