staging: media: wave5: vpu: Add support for hrtimer based poll
Add support for starting a polling timer in case interrupt
is not available. This helps keep the VPU functional in
cases where the hardware interrupt hookup may not be present.

Run hrtimer callback with 5ms polling interval while
any of the encoder/decoder instances are running to check
the interrupt status as being done in irq handler.

Based on above interrupt status, use a worker thread to
iterate over the interrupt status for each instance
and send completion event as being done in irq thread function.

Protect hrtimer access with device specific mutex locks
as this is shared across multiple instances of encoder
and decoder.

Add module param to change polling interval for debug purpose.

Co-developed-by: Sekhar Nori <nsekhar@ti.com>
Signed-off-by: Sekhar Nori <nsekhar@ti.com>
Signed-off-by: Devarsh Thakkar <devarsht@ti.com>