media: cadence: csi2rx: Fix YUV422 formats
For CSI-2 the correct formats to use for YUV422 are of the form 1x16 as
it is a serial protocol. The formats of the form 2x8 are used for
parallel protocols like DVP.

Add support for the 1x16 formats and make it the default. We still keep
around the 2x8 formats to avoid any regressions.

Signed-off-by: Jai Luthra <j-luthra@ti.com>
Signed-off-by: Vignesh Raghavendra <vigneshr@ti.com>