HACK: arm64: dts: ti: k3-j784s4-evm: Fix boot
This is a temporary solution to fix the boot issue on
k3-j784s4-evm to unblock integration.

Disabling main_cpsw1 fixes boot.
The mcu_cpsw is still enabled and is used for boot/NFS
so it should be sufficient for Integration.

The patch will be reverted when root cause is fixed.

Signed-off-by: Roger Quadros <rogerq@kernel.org>
Signed-off-by: Vignesh Raghavendra <vigneshr@ti.com>