arm64: dts: k3-am62a-main: Use all 6 CSI-RX channels
For multi-camera usecases we need all 6 capture channels to be
available.

Specifically for 4-stream usecases, like 2x OV2312 or 4x IM390 or
similar, we need more than 4 channels, as channel "rx0" can cause an
issue - where frames from other streams (VC != 0) end up on that rx0 if
the channel for that particular stream is not configured in time.

Signed-off-by: Jai Luthra <j-luthra@ti.com>
Signed-off-by: Vignesh Raghavendra <vigneshr@ti.com>