Revert "HACK: arm64: dts: ti: k3-j784s4-evm: Fix boot"
This reverts commit aca2f070328eeedbd02d03144c5782f5862bfbb3.

The am65_cpts_disable() function attempts to write to the CPTS registers
in order to reset the module. However, since the am65_cpts_release()
action which invokes am65_cpts_disable() is a devm action, in the event
of probe failure or module removal, the CPTS hardware is powered off
before the am65_cpts_disable() function can write to CPTS registers.
This is the root cause of the boot issue on J784S4 EVM.

Since this issue is fixed by:
"net: ethernet: ti: am65-cpsw/cpts: Fix CPTS release action",
enable the main_cpsw1 node in the device-tree for J784S4 EVM.

Signed-off-by: Siddharth Vadapalli <s-vadapalli@ti.com>
Signed-off-by: Vignesh Raghavendra <vigneshr@ti.com>