arm64: dts: ti: k3-am625-sk-lpmdemo: Enable audio
McASP1 and TLV320AIC3X (codec) are used for audio playback & recording
on SK-AM62x. Enable them both as basic suspend/resume functionality
works.

Active use-case is still under debug - if the system is suspended while
playback or recording is on-going and resumed later on, it does not
resume playback or recording from where it left with the currently
tested userspace applications (aplay/arecord/gst-play).

Signed-off-by: Jai Luthra <j-luthra@ti.com>
Tested-by: Dhruva Gole <d-gole@ti.com>
Signed-off-by: Vignesh Raghavendra <vigneshr@ti.com>