arm64: dts: ti: k3-j721s2-cpb-ov5640: Use powerdown instead of reset
The LI OV5640 module exposes single GPIO pin named RESETB, but the behavior
of the RESET signal is similar to powerdown signal expected by driver
and seems to be wired through appropriate inverters and delays internally
to PWDN and then to RESETB of the sensor.

For correct power-up sequence in the driver, we should pass this pin as
an active-low powerdown in the DT instead of calling it reset, otherwise
probe fails. Also add a comment to specify that this overlay is for LI
OV5640 module.

Signed-off-by: Vaishnav Achath <vaishnav.a@ti.com>
Reviewed-by: Siddharth Vadapalli <s-vadapalli@ti.com>
Reviewed-by: Jai Luthra <j-luthra@ti.com>
Signed-off-by: Vignesh Raghavendra <vigneshr@ti.com>