firmware: ti_sci: Use dt provided fw name to load at suspend time
Upstream: https://lore.kernel.org/all/20221128140522.49474-6-g-vlaev@ti.com/

Use request_firmware_direct to load the fs stub LPM firmware to a
provided memory region. The filename for the firmware is provided
in the device tree as "firmware-name".

Use a pm_notifier for loading of the low power firmware during
PM_SUSPEND_PREPARE phase so that it can be loaded from the rootfs
before it is suspended. It is possible in the future for this
firmware to require reload, so add a check to indicate that the
firmware is currently loaded so it is only loaded once.

Signed-off-by: Georgi Vlaev <g-vlaev@ti.com>
[backport: Selected relevant hunks for ti-linux-5.10.y]
Signed-off-by: Vibhore Vardhan <vibhore@ti.com>