remoteproc/omap: Add support for graceful shutdown
This patch adds support for graceful shutdown of the TI remoteproc
driver on a stop command. On some TI SoCs, the remote processor
(ex: M4F on AM64/AM62 SoCs) needs to be in quiescent state, before a
reset is asserted. This patch introduces a mailbox handshake mechanism
between the MPU and the remote processor to allow the remote processor
to gracefully shutdown its peripherals and enter a quiescent state.

Signed-off-by: Hari Nagalla <hnagalla@ti.com>