remoteproc: k3-m4: extend stop to send shutdown message to m4 core
This patch adds support for graceful shutdown of the K3 M4F remoteproc
driver on a stop command. On K3 SoCs, the remote processor needs to be
in a quiescent state, before a reset is asserted by the remoteproc
driver. This patch introduces a mailbox handshake mechanism between the
MPU and the remote processor to allow the remote processor to gracefully
shutdown its peripheral drivers and applications before entering a
quiescent state, which is typically entering 'wfi' with interrupts
disabled.

Signed-off-by: Hari Nagalla <hnagalla@ti.com>