dt-bindings: ti, sci: Add lpm region and firmware-name
Add documentation for the lpm region which tells the ti-sci driver where
to load the FS Stub low power mode firmware and also the firmware-name
which tells the driver which binary to load. Both of these are optional
for normal system operation but required to enabled suspend-to-mem usage
of Deep Sleep state.

Signed-off-by: Dave Gerlach <d-gerlach@ti.com>
Signed-off-by: Vibhore Vardhan <vibhore@ti.com>
Signed-off-by: Georgi Vlaev <g-vlaev@ti.com>
Reviewed-by: Rob Herring <robh@kernel.org>
Reviewed-by: Vibhore Vardhan <vibhore@ti.com>
Reviewed-by: Andrew Davis <afd@ti.com>
Link: https://serenity.dal.design.ti.com/lore/linux-patch-review/20230413121227.35114-2-d-gole@ti.com