media: ti: j721e-csi2rx: Set the data size shift correctly
The SIZE_CFG field in SHIM_DMACNTX field needs to be set correctly for
the image to be unpacked correctly. Currently, the field is set to the
default value of 0 which works fine for YUV422-8 and RAW6-8. But for other
formats like RAW10-16 or YUV422-10 it results in the second byte of the
pixel being dropped.

Introduce a new member to ti_csi2rx_fmt that stores the size and set it
when configuring the DMA context to make sure data is unpacked
correctly.

Signed-off-by: Pratyush Yadav <p.yadav@ti.com>
Signed-off-by: Jai Luthra <j-luthra@ti.com>