media: platform: img: vxe-vxd: fix a array out of bounds bug
This changed sometime after kernel 4.19.  One reserved field was
removed, so writing to the v4l2_ext_controls reserved field for
array index 0 and 1 actually is an array out of bounds write.

I am not entirely sure why we are writing to a reserved field in
the first place.  But I'll fix this bug for now, and revisit that
question at a later date.

Signed-off-by: Darren Etheridge <detheridge@ti.com>