dt-bindings: gpu: pvr: combine rogue devices
Clean this bindings file up and combine the rogue devices now that J*
hybrid APM is up and running.

Drop the power island pd on BXS/8XE. The power island is controlled
directly by the firmware through the use of an MMR on all of our
platforms. Registering this domain with the Linux power system
interferes with this behavior.

Now all rogue devices behave the same and require the same dt
attributes.

Signed-off-by: Randolph Sapp <rs@ti.com>