arm64: dts: ti: k3-j784s4-common-proc-board: Enable main_i2c5 and disable csi ports
Enable main_i2c5 instance which is the camera control I2C interface
for the CSI2 instances on J784S4 and also enable the TCA6408 GPIO
expander on main_i2c5. CSI2RX instance 0 and instance 1  are routed
to the Samtec MIPI connector 0 and CSI2RX instance is routed to the
MIPI connector 1 (CSI2-EXP-AUX) connector  on the EVM thus enable
them as well.

Signed-off-by: Vaishnav Achath <vaishnav.a@ti.com>
Signed-off-by: Vignesh Raghavendra <vigneshr@ti.com>