arm64: dts: ti: k3-am68-sk: Add DT node for USB
AM68 Starter kit has a USB3 hub that connects to the SerDes0 Lane 2.
Add DT node to link the USB3 hub to the SerDes USB PHY.

Also, the SerDes driver can only support two protocols at a time. In the
j721s2_main.dtsi, the DP PHY is enabled by default. As PCIe and USB3
are connected to SerDes0 in the AM68 SK, disable the unused DP PHY node.

Signed-off-by: Sinthu Raja <sinthu.raja@ti.com>
Reviewed-by: Ravi Gunasekaran <r-gunasekaran@ti.com>
Signed-off-by: Vignesh Raghavendra <vigneshr@ti.com>