arm64: defconfig: Make Serdes and WIZ drivers built-in
The Serdes drivers (Sierra and Torrent) and the WIZ driver are currently
built as modules. The client drivers which use the Serdes are built-in.
Due to this, it is possible that the long delay in the Serdes drivers
being probed will result in a probe deferral of the client drivers as
well.

Prevent this by having the Serdes and WIZ drivers built-in.

Signed-off-by: Siddharth Vadapalli <s-vadapalli@ti.com>
Acked-by: Ravi Gunasekaran <r-gunasekaran@ti.com>
Signed-off-by: Vignesh Raghavendra <vigneshr@ti.com>