arm64: dts: ti: k3-am62a-main.dtsi: Use SRAM for VPU
The wave5 VPU has a separate AXI interface to on chip SRAM
which can be used for faster access to internal data being
used for processing.

Allow VPU to allocate 62KiB of SRAM memory so that it can
use the same for storing and accessing temporal buffers
while any of the VPU instance is running.

62KiB is chosen since AVC decoder requires 62KiB
for 4K resolution, although HEVC requires 65Kib
but since we only have total 64KiB available,
going with 62Kib.

Signed-off-by: Devarsh Thakkar <devarsht@ti.com>
Acked-by: Prasanth Babu Mantena <p-mantena@ti.com>
Signed-off-by: Vignesh Raghavendra <vigneshr@ti.com>