arm64: dts: ti: Add overlay for MCAN in AM62x
On AM62x there is 1x MCAN in MAIN domain and 2x MCAN in MCU domain.

There is no hardware interrupts routed to A53 GIC interrupt controller
but timer polling was introduced to MCAN in [1] and in this series MCU
MCAN nodes are added to device tree.

AM62x does not carry on-board CAN transceivers, so MCAN does not work
'out of the box', external CAN transceiver need to be connected. Instead
of changing DTB permanently use an overlay to enable 3x MCANs and for
CAN transceiver nodes.

If there is no hardware interrupt and timer polling method is to be used,
do not include interrupt properties, this enables polling method on MCAN.

[1] Upstream commit shaw: 7ba402035eea711c5ae268630c4bc3dc9afe8645

Signed-off-by: Judith Mendez <jm@ti.com>
Signed-off-by: Vignesh Raghavendra <vigneshr@ti.com>