arm64: dts: k3-am69-sk: Add overlay for DDR carveout memory nodes
Two carveout reserved memory nodes each have been added for each of the
R5F and C71x remote processor devices/DSP/DSPs within both the MCU and
MAIN domains. These nodes are assigned to the respective rproc device
nodes as well. The first region will be used as the DMA pool for the
rproc device, and the second region will furnish the static carveout
regions for the firmware memory. These should match with the IPC and
external memory sections of the remote processor FW binary.

Signed-off-by: Dasnavis Sabiya <sabiya.d@ti.com>
Signed-off-by: Vignesh Raghavendra <vigneshr@ti.com>