arm64: dts: ti: Add J7200 MCSPI loopback overlay
J7200 has main_mcspi4 internally connected as slave to mcu_spi2 master,
this commit adds an overlay for demonstrating the internal MCSPI loopback
feature and acts a reference for enabling DMA and slave mode for MCSPI.
MCSPI slave mode is supported only with DMA enabled. The same overlay can
be used for other J7 platforms with update in the PSIL thread ID, the
necessary information will be added in documentation.

Signed-off-by: Vaishnav Achath <vaishnav.a@ti.com>
Reviewed-by: Dhruva Gole <d-gole@ti.com>
Reviewed-by: Dhruva Gole <d-gole@ti.com
Signed-off-by: Vignesh Raghavendra <vigneshr@ti.com>