remoteproc: k3-m4: Enable MCU wakeup support
Add support for MCU (m4f) wakeup. When MCU wakeup is enabled
(/sys/bus/platform/devices/xxxx/power/wakeup), on a system suspend
a IPC message (SYSTEM_SUSPEND) is sent to the remote core and waits
until an ACK is received. MCU would be in active power state when
MCU wakeup is enabled.

[d-gole@ti.com] minor checkpatch warnings addressed

Signed-off-by: Hari Nagalla <hnagalla@ti.com>
Signed-off-by: Dhruva Gole <d-gole@ti.com>