arm64: dts: ti: Add dmtimer-pwm overlay for AM62x
Based on the upstream patch by Tony Lindgren [1].

On am625, pad c17 can be used either in gpio mode or pwm mode to control
the connected led. Let's change to use pwm-led for additional brighness
control.

The pwm controller is dmtimer7 for pad C17. We configure it for 128 Hz
rate (1000000000 / 7812500) similar to what has been done elsewhere for
pwm-leds.

[1] https://lore.kernel.org/all/20221107163533.48777-5-tony@atomide.com/

Signed-off-by: Vibhore Vardhan <vibhore@ti.com>
Signed-off-by: Vignesh Raghavendra <vigneshr@ti.com>