arm64: defconfig: Enable RPMSG_KDRV_ETH_SWITCH config
The rpmsg_kdrv_switch.c driver is used by the j721e-cpsw-virt-mac.c
driver for binding to the remote core running Eth Switch Firmware.

Thus, enable the config for the rpmsg_kdrv_switch.c driver to be
built-in. Building it as a module will result in the
j721e-cpsw-virt-mac.c driver's probe being deferred due to its
dependency on the rpmsg_kdrv_switch.c driver.

Signed-off-by: Siddharth Vadapalli <s-vadapalli@ti.com>
Signed-off-by: Vignesh Raghavendra <vigneshr@ti.com>