net: ethernet: ti: icssg_prueth: Flush FDB entries only in MAC mode
In Switch and HSR mode, FDB table is updated in ndo_open(). But
commit b899c9072093 ("net: ti: ethernet: icssg_prueth: Add Multicast
filtering support in MAC mode.") flushes all the FDB entries in rx_mode().
And this affects the Rx functionality.

So flush the FDB entries and update VLAN table in rx_mode() only in MAC
mode for now.

Signed-off-by: Ravi Gunasekaran <r-gunasekaran@ti.com>
Signed-off-by: Vignesh Raghavendra <vigneshr@ti.com>