net: ethernet: ti: icssg_prueth: Enable IEP1 in HSR mode
commit bfae54d64b1f ("net: ethernet: ti: icssg_prueth:
Add support for HSR frame forward offload") introduced HSR offload
support with a known issue of brief stalling with small packet sizes.

In HSR firmware, IEP1 is used to periodically age out entries (about
every 10ms and entryForget time is 400ms) so that valid entries
are not discarded while forwarding. If IEP1 is not enabled, it will
not ageout entries and firmware may make incorrect packet drops.

So enable and disable IEP1 when entering and exiting HSR mode.

Fixes: bfae54d64b1f ("net: ethernet: ti: icssg_prueth: Add support for HSR frame forward offload")
Signed-off-by: Ravi Gunasekaran <r-gunasekaran@ti.com>
Reviewed-by: Roger Quadros <rogerq@kernel.org>
Signed-off-by: Vignesh Raghavendra <vigneshr@ti.com>