arm64: dts: ti: Add ECAP Capture overlay for AM62x
Add an overlay to test ECAP Capture mode for AM62x SK. All ECAP modules
are configured in PWM mode by default. This commit replaces the default
properties for ECAP2 module to enable the capture mode. It also
configures the padconf to capture input signal on pin 11 of J3 connector.

Signed-off-by: Vibhore Vardhan <vibhore@ti.com>
Signed-off-by: Vignesh Raghavendra <vigneshr@ti.com>