media: platform: img: e5010: Remove job_ready callback
Remove job_ready callback since v4l2 m2m core already
ensures to queue the job only if non-zero buffers are available
at capture and output queue and that criteria suffices the driver.

This also removes the check for hardware busy flag as it was causing
a hang issue in multistream scenario where userspace queues the buffer
which triggers the job_ready callback and the job is not queued
ever since the hardware was busy and userspace has queued all the
buffers.

The hardware busy check as such is not required for queuing the job
since the e5010_hw* functions already take care of polling the busy
flag and ensuring that registers are only programmed when hardware is not
busy, hence remove the now unused function e5010_hw_is_busy too.

While at it, remove extra traversal for e5010_dev struct wherever
possible for dynamic debug prints.

Fixes: 995721196ec4 ("media: platform: img: Add E5010 JPEG Encoder driver")

Signed-off-by: Devarsh Thakkar <devarsht@ti.com>
Reviewed-by: Darren Etheridge <detheridge@ti.com>