drm/bridge: ti-sn65dsi86: Use drm_connector_init with helper funcs
In the current driver, the bridge detect hook is not called due to which
the DP1 connector connected to the datapath for dsi (TIDSS->CDNS-DSI->
SN65DSI86->DP1-connector) always comes up as connected. So to fix this,
now the connector is initialized by drm_connector_init call with bridge
connector_funcs instead of drm_bridge_connector_init call.
Another member "plugged" is added to the platform data for ti_sn65dsi86
structure which indicates if the panel is plugged or not by reading the
dpcd_link_status in bridge attach call. Based on the value of "plugged",
the detect hook of bridge_conenctor_funcs return the connector status.

Signed-off-by: Jayesh Choudhary <j-choudhary@ti.com>
Reviewed-by: Rahul T R <r-ravikumar@ti.com>
Reviewed-by: Aradhya Bhatia <a-bhatia1@ti.com>