HACK: arm64: configs: pcie: enable PCIe on AM65 platforms
To get the PCIe devices(NVME SSD, WIFI etc) working with out of box
TI Linux SDK, enabling PCIe driver for TI's AM65 SOC to be build as
a built-in kernel driver.

pci-keystone driver is not made in a way to be a build as a kernel loadable
module, so it has to be enabled as 'y' and it depends on am654-serdes phy
driver which also has to enabled as 'y' else absence of phy driver puts
pci-keystone for deferred probe and in later point of time when
pci-keystone probe() is called again, we get crash due to pointer
dereference fault as section that contains probe() is marked __init.

Signed-off-by: Achal Verma <a-verma1@ti.com>