usb: cdns3: Set default EP buffer size
commit dce49449e04f ("usb: cdns3: allocate TX FIFO size according to
composite EP number") changed the ep_buf_size from a default value of
4 and introduced a mechanism to calculate the buffer size based on the
IN EPs per configuration. This calculation is carried out only for the
gadgets created using configfs and breaks for legacy gadget drivers
for CDNS3 UDC.

So start the UDC driver with the previously known safe ep_buf_size,
so that it works for gadget driver that don't invoke usb_gadget_check_config().

Fixes: dce49449e04f ("usb: cdns3: allocate TX FIFO size according to composite EP number")
Signed-off-by: Ravi Gunasekaran <r-gunasekaran@ti.com>