arm64: dts: ti: k3-am68-sk-imx219: Fix CSI mux selection for RPi camera
The AM68 SK board includes a multiplexer for selecting between the CSI MIPI
(Port A) and FPC connectors (Port B). The FPC connector (Port B) is used to
attach the RPi camera. The CSI data from the RPi camera did not reach the
CSI interface since Port A was set by default. To enable the port B for the
RPi camera, pull the CSI_MUX_SEL line to high.

Fixes: 78e0aa2bed4e ("arm64: dts: ti: k3-am68-sk: Add DT overlay for RPi camera V2")
Signed-off-by: Sinthu Raja <sinthu.raja@ti.com>