drm/tidss: Make zpos immutable for primary plane
The new version of weston seems to be setting
non-zero value of zpos for primary plane too [1].

As primary plane sits at the bottom and is always
"on" and overlay planes are "overlayed" on top
of primary, set zpos value of primary plane as
immutable with 0 value.

For overlay plane, set minimum value in range for zpos
value as 1 so that it doesn't overlap with
value for primary plane.

This is based off :
https://lore.kernel.org/linux-renesas-soc/20200401061100.7379-1-etom@igel.co.jp/
https://patchwork.kernel.org/project/linux-renesas-soc/patch/20200402104035.13497-1-laurent.pinchart+renesas@ideasonboard.com/

[1]: http://serenity.dal.design.ti.com:7777/ayoviboceg.yaml

Signed-off-by: Devarsh Thakkar <devarsht@ti.com>
Signed-off-by: Vignesh Raghavendra <vigneshr@ti.com>