media: platform: img: vxd fix the mapping problem with dmabuf import
If a position seek is performed within GStreamer, the MMU mappings are
destroyed.  However the flag to indicate that the buffers are mapped is
still set.  Therefore the driver blindly assumes that the mapping is
still valid when it tries to pickup the playback from the next position.
Whereas it should be remapping the buffers into the MMU.  This only
happens during seek, because the buffers get reused instead of
completely torn down as is the case with a full stream shutdown.

Signed-off-by: Darren Etheridge <detheridge@ti.com>
Signed-off-by: Vignesh Raghavendra <vigneshr@ti.com>