Revert "drm/tidss: Make zpos immutable for primary plane"
This reverts commit e0757b737833c0baee39cf54de96abb15a7b4a24.

Fixing zpos to 0 for two primary planes causes issue when one
video pipeline tries to use all other planes as overlay as we
cannot change the zpos higher than background.
This effectively means that we can use only two instead of three
planes as overlay (for J7 devices with 4 OVR and VP).
Therefore reverting the patch.

Signed-off-by: Jayesh Choudhary <j-choudhary@ti.com>
Reviewed-by: Aradhya Bhatia <a-bhatia1@ti.com>