spi: omap2-mcspi: Add FIFO support without DMA
commit 75223bbea840e125359fc63942b5f93462b474c6 upstream.

Currently, the built-in 64-byte FIFO on the MCSPI controller is not
enabled in PIO mode and is used only when DMA is enabled. Enable the
FIFO in PIO mode by default and fallback only if FIFO is not available.
When DMA is not enabled, it is efficient to enable the RX FIFO almost
full and TX FIFO almost empty events after each FIFO fill instead of
each word. Update omap2_mcspi_set_fifo() to enable the events accordingly
and also rely on OMAP2_MCSPI_CHSTAT_RXS for the last transfer instead of
the FIFO events to handle the case when the transfer size is not a
multiple of FIFO depth.

See J721E Technical Reference Manual (SPRUI1C), section 12.1.5
for further details: http://www.ti.com/lit/pdf/spruil1

Link: https://lore.kernel.org/r/20231013092629.19005-1-vaishnav.a@ti.com
Signed-off-by: Mark Brown <broonie@kernel.org>
Signed-off-by: Vaishnav Achath <vaishnav.a@ti.com>