counter/ti-eqep: implement over/underflow events
This adds support to the TI eQEP counter driver for subscribing to
overflow and underflow events using the counter chrdev interface.

Since this is the first event added, this involved adding an irq
handler. Also, additional range checks had to be added to the ceiling
attribute to avoid infinite interrupts.

Link: https://www.spinics.net/lists/linux-iio/msg64814.html

Signed-off-by: David Lechner <david@lechnology.com>
[Judith: Fix merge conflicts and grammar errors. Only clear interrupts
that were addressed. Add link to commit message]
Signed-off-by: Judith Mendez <jm@ti.com>