counter/ti-eqep: add support for edge capture unit
This adds support for the Edge Capture Unit to the TI eQEP counter
driver. This just adds the minimum required features to measure speed
using the Unit Timer and the Edge Capture unit. Additional features can
be added in the future if needed.

This adds 4 new device-level attributes:
- edge_capture_unit_prescaler: selects a prescalar for the Counter count
	coming into the Edge Capture Unit
- edge_capture_unit_max_period: selects the max time period that can be
	measured by the Edge Capture Unit
- edge_capture_unit_latched_period: gets the period that was measured
	when the event selected by the latch_mode attribute is triggered
- edge_capture_unit_enable: enables or disables the Edge Capture Unit

Link: https://www.spinics.net/lists/linux-iio/msg64577.html

Signed-off-by: David Lechner <david@lechnology.com>
[Judith: Fix merge conflicts. Add link to commit message. Change clock
to "fck"]
Signed-off-by: Judith Mendez <jm@ti.com>