net: ethernet: ti: Add CPSW Proxy Client driver
The CPSW Proxy Client driver in Linux communicates with the Ethernet
Switch Firmware (EthFw) running on a remote MAIN R5 core via RPMsg.
EthFw announces its endpoint over the RPMsg-Bus, triggering the probe
of the CPSW Proxy Client driver. Following this, the communication
between CPSW Proxy Client driver and EthFw can proceed over RPMsg and
the Client driver can request various services provided by EthFw which
facilitate the transmission and reception of Network Data by the Client.

EthFw is in complete control of the CPSWnG peripheral and any configuration
pertaining to the CPSWnG peripheral can only be performed by EthFw at the
request of the Client. This model allows the shared use of the CPSWnG
peripheral between the clients running on different cores.

The data is transmitted/received over the DMA Channels which EthFw
allocates to the Client. The Client configures the DMA Channels
allocated to it using the DMA APIs provided by the DMA driver.

Signed-off-by: Siddharth Vadapalli <s-vadapalli@ti.com>
Reviewed-by: Jai Luthra <j-luthra@ti.com>
Tested-by: MD Danish Anwar <danishanwar@ti.com>