arm64: defconfig: Enable CPSW Proxy Client driver
The CPSW Proxy Client driver communicates with the Ethernet Switch
Firmware (EthFw) running on a remote MAIN R5 core via RPMsg. EthFw
configures the CPSWnG instance of the CPSW Ethernet Switch on TI's
SoCs and provides services to the clients running on different cores
of which the CPSW Proxy Client is the client on the core running
Linux.

Enable the CPSW Proxy Client driver as a module for availing the
services provided by EthFw.

Signed-off-by: Siddharth Vadapalli <s-vadapalli@ti.com>
Reviewed-by: Jai Luthra <j-luthra@ti.com>
Tested-by: MD Danish Anwar <danishanwar@ti.com>