arm64: defconfig: Modify number of usable 8250 uart ports
Jacinto SoCs have a total of 12 serial uart ports. But by default,
only 4 out of these 12 uarts can be used at a time instead which is
default in Kconfig.

Thus, Modify maximum number of 8250 serial uart ports which can be
used as well as modify how many 8250 serial uart ports which we use
in runtime to 12 for all the SoCs.

Signed-off-by: Bhavya Kapoor <b-kapoor@ti.com>