can: m_can: Map WoL to device_set_wakeup_enable
In some devices the m_can module can act as a wakeup source. This patch
helps do that by connecting the PHY_WAKE WoL option to
device_set_wakeup_enable. By marking this device as being wakeup
enabled, this setting can be used to set appropriate options to enable
this device as a wakeup source.

Signed-off-by: Markus Schneider-Pargmann <msp@baylibre.com>