media: ti: j721e-csi2rx: Re-use a 32KiB drain-buffer
Only allocate 32KiB of memory for the drain buffer, which is the upper
bound of the PSI-L endpoint's internal FIFO storage (2048 x 128-bits)

Also, allocate the drain buffer only once in probe() instead of
start_streaming() for every context. As we don't care about the drained
data, we can flush it out to the same buffer for different streams.

Signed-off-by: Jai Luthra <j-luthra@ti.com>