net: ti: icssg-prueth: Fix TX buffer pool size in switch mode
Currently the buffer size for tx queues in switch mode is 2KB. Due to
this larger packet (MTU size) ping fails. To fix this the buffer size
per queues needs to be changed to 4KB. However this results into
increase in MSMC ram needed for ICSSG. Fix this by reducing the number
of queues per port from 16 to 8 in switch mode as the driver only uses 8
queues per port. Firmware supports 16 queues per port (8 for pru0 cores
and 8 for pru1 cores) in switch mode but the driver only uses 8 queue per
port (4 for pru0 cores and 4 for pru1 cores). Allocate only the queues
being used by driver.

Now the MSMC requirement remains the same and the buffer size is increased
to support ping for large size packets.

Fixes: 2623e9589873 ("net: ethernet: ti: icssg_prueth: Add support for ICSSG switch firmware on AM654 PG2.0 and AM64 EVM")
Signed-off-by: MD Danish Anwar <danishanwar@ti.com>
Tested-by: Ravi Gunasekaran <r-gunasekaran@ti.com>