arm64: dts: k3-am62p: Enable HS400 speed mode
HS400 speed mode depends on High speed and HS200 speed modes
so add these to the eMMC node.

The SR12-SR104 and DDR50 speed modes are not working, remove
these speed modes.

Characterization is not complete for HS, HS200, and HS400 speed
modes, but the process should be to use the base otap value and
update in the future if characterization shows otherwise. Use otap
6 for HS200/HS400. For HS, use the itap/otap based off of Jacinto
device tree.

Signed-off-by: Judith Mendez <jm@ti.com>